config: simplify default for local testing
Remove the the sals/NickServ-authenticated parts from defualt.json because
these are difficult to unset for local testing, due to the deep-merge
that node-config performs.

It's much easier to add these properties in production, then to unset
them in local testing.

Keep the sample through a an example-production.json file instead.

Change-Id: I33027397e2b3ec9c8a93950151ae664cb3ee8a40